renamed seq_match to fuzzy
